Usage

BijectiveHilbert.brgc

BijectiveHilbert.brgc_inv

BijectiveHilbert.brgc_inv

BijectiveHilbert.decode_hilbert

BijectiveHilbert.decode_hilbert!

BijectiveHilbert.decode_hilbert_zero

BijectiveHilbert.encode_hilbert

BijectiveHilbert.encode_hilbert

BijectiveHilbert.encode_hilbert_zero

BijectiveHilbert.hi_d

BijectiveHilbert.hi_f

BijectiveHilbert.hilbert_order

BijectiveHilbert.brgc — Method

Binary-reflected Gray code.

BijectiveHilbert.brgc_inv — Method

This is how the paper describes the inverse of the binary-reflected gray code.

BijectiveHilbert.brgc_inv — Method

Inverse of the binary-reflected Gray code. This takes you from the Gray code to its index.

BijectiveHilbert.decode_hilbert! — Method

decode_hilbert(x, y)

A 1-based Hilbert decode, from decode_hilbert_zero.

BijectiveHilbert.decode_hilbert — Method

decode_hilbert(x, y)

A 1-based Hilbert decode, from decode_hilbert_zero.

BijectiveHilbert.decode_hilbert_zero — Method

decode_hilbert_zero(z::Integer) -> (x, y)

Computes the (x, y) from a Hilbert code.

This function is zero-based. 0 <= x < 2^n, 0 <= y < 2^n, and 0 <= z < 4^n.

See also: encode_hilbert_zero, decode_hilbert.

BijectiveHilbert.encode_hilbert — Method

encode_hilbert(x, y)

A 1-based Hilbert code, so x and y start at one, and the z that this returns also starts at one.

See also: encode_hilbert_zero.

BijectiveHilbert.encode_hilbert — Method

encode_hilbert(gg::HilbertAlgorithm{T}, X::Vector{A})

A 1-based Hilbert encoding.

BijectiveHilbert.encode_hilbert_zero — Method

encode_hilbert_zero(x::Integer, y::Integer)

Computes an integer Hilbert index for x and y using a variant algorithm.

Given two integer indices for a 2-dimensional plane, return a single index. This index is designed to increase locality for 1-dimensional access. It does this by keeping nearby points in 2 dimensions also nearby in 1 dimension.

x and y need to be integers that have bit-shifting operations.

The variant algorithm used differs from the usual Hilbert code because it doesn't need to know the size of the whole grid before computing the code [1]. It looks like a slightly-rotated version of the Hilbert curve, but it has the benefit that it is 1-1 between (x, y) and z, so you can translate back and forth.

This function is zero-based. 0 <= x < 2^n, 0 <= y < 2^n, and the result is 0 <= z < 4^n.

See also: decode_hilbert_zero, encode_hilbert.

BijectiveHilbert.hi_d — Method

page 12. directions d(i) = 0 if i=0 = g(i-1) mod n if i=0 mod 2 = g(i) mod n if i=1 mod 2. Domain is 0 <= i <= 2^n - 1.

BijectiveHilbert.hi_f — Method

f is the exit vertex of the ith sub-hypercube in a Gray code ordering of the sub-hypercubes. Corllary 2.7 on pg. 12 says: hif(i, n) = hie(i) ⊻ hi_d(i, n)

BijectiveHilbert.hilbert_order — Method

hilbert_order(v::AbstractArray, subdivisions)

Calculates the permutation of v that would order a 2D array by Hilbert curve, where v is an array of real numbers, dimension (2, N) and subdivisions is a real number that specifies how many boxes to place the v into, per side.

This tells you how to order an array of 2-dimensional numbers so that they have more memory locality in 1 dimension. Given an array of real numbers of dimension (2, n), subdivide them in each dimension by subdivisions, and assign each point a Hilbert code. Return the permutation that would sort the given array by that Hilbert code.

See also: encode_hilbert_zero.

Example

rng = MersenneTwister(984720987)
points_in_space = zeros(2, 100)
rand!(rng, points_in_space)
points_reordered = points_in_space[:, hilbert_order(points_in_space, 50)]

1N. Chen, N. Wang, B. Shi, A new algorithm for encoding and decoding the Hilbert order. Software—Practice and Experience 2007; 37(8): 897–908.